Revert "Convert `source-actors` reducer to TS (#6209)"
This reverts commit 0b77cb51b88cfd43a578850f3c3974ea24c67c24.
